+++ changelog.md
@@ -4,6 +4,31 @@ All the notable additions and fixes.
 
 This changelog follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+# [1.1.5] - 2024-12-27
+
+### Changed
+
+- Updated login form to use react-hook-form
+
+### Fixed
+
+- Compile error due to width and height not specified for Image component in account page
+
+# [1.1.4] - 2024-12-26
+
+### Added
+
+- /reset-password route along with `ResetPasswordFormData` type
+- Email notifications for resetting password alongside new email templates
+- `reset_password` database table. Schema added to notion
+- Token generation, storing, and clearing from redis instance
+- All appropriate validation for token (expired, invalid, multi-requested etc.)
+- `api.md` for API reference. Filled with new updates **only** for push 1.1.4
+
+### Changed
+
+- useCallback for useEffect dependencies inside existing codebase (as per build warnings)
+
 # [1.1.3] - 2024-12-24
 
 ### Added
@@ -80,5 +105,4 @@ This changelog follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Contact Us page
 - Admin page (tables)
 - Home page (carousel, snap scroll)
-